+++ opennmt/models/sequence_to_sequence.py
@@ -131,12 +131,13 @@ class SequenceToSequence(Model):
     self.alignment_file_key = alignment_file_key
     self.alignment_file = None
 
-  def _initialize(self, metadata):
-    super(SequenceToSequence, self)._initialize(metadata)
+  def _initialize(self, metadata, asset_dir=None):
+    assets = super(SequenceToSequence, self)._initialize(metadata, asset_dir=asset_dir)
     if self.alignment_file_key is not None and self.alignment_file_key in metadata:
       self.alignment_file = metadata[self.alignment_file_key]
       tf.logging.info(
           "Training with guided alignment using alignments from %s", self.alignment_file)
+    return assets
 
   def _augment_parallel_dataset(self, dataset, process_fn, mode=None):
     # Possibly add alignments as labels.